+++ docs/components/FooterNotes.vue
@@ -1,12 +1,23 @@
 <template>
-	<div class="border-t border-alpha-1 border-solid text-word-3 text-14 p-16">
-		proudly powered by
+	<div
+		class="border-t border-alpha-1 border-solid text-word-2 text-14 p-16 leading-relaxed">
+		<div>
+			Released under the
+			<a
+				href="https://opensource.org/license/mit/"
+				rel="noopener noreferrer"
+				target="_blank"
+				class="vv-button vv-button--link"
+				>MIT License</a
+			>.
+		</div>
+		Maintained with <span class="text-10">♥️</span> by
 		<a
-			href="https://24consulting.it"
+			href="https://8wave.it"
 			rel="noopener noreferrer"
 			target="_blank"
-			class="font-semibold"
-			>24/Consulting</a
+			class="vv-button vv-button--link"
+			>8 Wave</a
 		>
 	</div>
 </template>