+++ docs/components/Logo.vue
@@ -1,0 +1,3 @@
+<template>
+	<img src="/volverjs.svg" />
+</template>