+++ stylelint.config.cjs
@@ -2,7 +2,6 @@ module.exports = {
 	extends: [
 		'stylelint-config-standard-scss',
 		'stylelint-config-recommended-vue/scss',
-		'stylelint-config-prettier',
 	],
 	rules: {
 		'property-no-unknown': null,